Rename Container and Transforms to Sprite-like classes
