+++ Zilon.Core/CDT.LAST.MonoGameClient/Engine/SpriteContainer.cs
@@ -1,0 +1,9 @@
+﻿namespace CDT.LIV.MonoGameClient.Engine
+{
+    /// <summary>
+    /// A container that can hold and draw internal entities with shared transformations.
+    /// </summary>
+    public class SpriteContainer : Renderable
+    {
+    }
+}
